"Game" class layers; Editor uses Framework's.
Because of how Nez works, Core references MonoGame. Not ideal.
