--- source/SharpForge.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace SharpForge.Core;
-
-public class Class1
-{
-
-}
