+++ source/SharpForge.Core/Game.cs
@@ -1,0 +1,6 @@
+﻿namespace SharpForge.Core;
+
+public class Game : SharpForge.Backend.Nezz.Game
+{
+
+}
